FINALLY added the ability to include additional dialog boxes in the
rooms. The initial dialog box is still set as normal. Additional dialog
boxes are declared in the map class. Added method showNextDB(), which is
used to trigger dialog boxes passed the initial box and can be used in
unique conditions for different rooms. I added a box that appears after
a player beats the first challenge cavern in room 4, telling them about
the mechanic of staying or leaving the caverns. I also added a boolean
called "immune" to health bar. This is set to true when the controls are
locked and prevents damage from being dealt to a player by external
forces.